refacto(withLocales): Convert into typescript
